Fixed syntax errors in examples and protection levels in library.

+++ src/Demiurge.cpp
@@ -3,7 +3,7 @@
 
 bool Demiurge::_initialized = 0;
 
-Demiurge::Demiurge(uint32_t sampleRate) {
+Demiurge::Demiurge(long sampleRate) {
 
 }
