+++ src/Demiurge.h
@@ -1,12 +1,14 @@
 #ifndef _DEMIURGE_DEMIURGE_H_
 #define _DEMIURGE_DEMIURGE_H_
 
-#include <cstdint>
 
+#include "Adsr.h"
 #include "AudioInPort.h"
 #include "AudioOutPort.h"
 #include "CvInPort.h"
+#include "ControlPair.h"
 #include "GatePort.h"
+#include "Inverter.h"
 #include "Mixer.h"
 #include "Pan.h"
 #include "Potentiometer.h"
@@ -39,7 +41,7 @@ public:
    }
 
 private:
-   Demiurge(uint32_t sampleRate);
+   Demiurge(long sampleRate);
    ~Demiurge();
 
    static bool _initialized;
